[Refactor] refactor variant type code
