+++ be/src/exprs/variant_functions.h
@@ -16,7 +16,7 @@
 
 #include "column/column.h"
 #include "common/statusor.h"
-#include "formats/parquet/variant.h"
+#include "util/variant.h"
 #include "function_helper.h"
 #include "types/logical_type.h"
