+++ be/src/types/variant_value.h
@@ -20,7 +20,7 @@
 
 #include "common/statusor.h"
 #include "fmt/format.h"
-#include "formats/parquet/variant.h"
+#include "util/variant.h"
 #include "util/slice.h"
 
 namespace starrocks {
