+++ be/src/util/CMakeLists.txt
@@ -116,10 +116,10 @@ set(UTIL_FILES
   lake_service_recoverable_stub.cpp
   byte_buffer.cpp
   variant_converter.cpp
-  variant_util.cpp
   jvm_metrics.cpp
   llm_cache.cpp
   llm_query_service.cpp
+  variant.cpp
 )
 
 ADD_BE_LIB(Util
