+++ be/test/exprs/variant_functions_test.cpp
@@ -29,7 +29,7 @@
 #include "common/status.h"
 #include "common/statusor.h"
 #include "exprs/mock_vectorized_expr.h"
-#include "formats/parquet/variant.h"
+#include "util/variant.h"
 #include "fs/fs.h"
 #include "gutil/casts.h"
 #include "runtime/runtime_state.h"
